NEW BUG ENCOUNTERED LETS GOOOOOOOO
When you enter and exit the pause menu repeatedly (never letting the pause menu going off-screen completely), buildings and procedural stuff stops appearing.
Look if it's checking the wrong pause varaible or if there's a condition that prevents the function from decreasing the time counter when the pause menu exists or the timescale is lower than 1.
I also need to learn to do more descriptive descriptions, like this. And comment more.