Level selection scrolls in PC
Warning: Weird frustrating Unity's internal scripts error occurrs when trying to scroll or click the Level Buttons in a Phone device. I'm trying to work on that.
Also: uncomprehensible NullReferenceExepcion occurrs when entering the options scene in Phone.
Fuck Unity this is the last project in which I use it